enable npm and maven prerelease version syntax
